--- src/utils/cn.ts
@@ -1,5 +1,0 @@
-const cn = (classNames: unknown[]) => {
-  return classNames.filter(Boolean).join(' ');
-};
-
-export default cn;
