+++ src/utils/index.ts
@@ -1,6 +1,2 @@
-import cn from './cn';
-import mapToObj from './mapToObj';
-import objToMap from './objToMap';
-
-export { cn, mapToObj, objToMap };
+export * from './common';
 export * from './localStorage';
