--- src/utils/objToMap.ts
@@ -1,5 +1,0 @@
-const objToMap = (obj: object) => {
-  const entries = Object.entries(obj);
-  return new Map(entries);
-};
-export default objToMap;
